Moved passwordhashing function to lib file and created login page
